refactor(matrix-bots): deduplicate shared bot logic into shared/lib
- Extract parseCommand, createMessageHandler, and createInviteHandler
  to shared/lib/ so they live in exactly one place.
- Parameterise handler and invite-handler via config.bot.mode (dm|room),
  config.bot.botType, and config.bot.fallbackMessageSuffix so per-bot
  differences are driven by config, not forked copies.
- Remove duplicated lib/ and tests/ from dmbot/ and roombot/.
- Add minimal wiring tests in dmbot/tests/ and roombot/tests/.
- Create comprehensive shared test suite (43 tests) covering both
  DM and room modes.
- Update Dockerfiles to copy shared/ into the image and use /app/<bot>
  as WORKDIR.
- Update docker-compose.yml to use repo root as build context.
- Update CI workflow to run shared tests from dmbot directory.
- Update README.md and CONTRIBUTING.md to reflect new test layout.

Closes #8
