+++ shared/lib/commands.js
