+++ shared/lib/handler.js
@@ -2,16 +2,40 @@ import { parseCommand } from "./commands.js";
 
 // Builds the room.message handler. Dependencies are injected so the handler
 // can be unit-tested without a live Matrix client or n8n instance.
+// Bot-specific behaviour is driven by config.bot.mode and related fields.
 export function createMessageHandler({ client, config, axios, botUserId }) {
+  const mode = config.bot.mode ?? (config.bot.allowedUsers ? 'dm' : config.bot.targetRoomId ? 'room' : 'dm');
+  const botType = config.bot.botType ?? mode;
+  const fallbackSuffix = config.bot.fallbackMessageSuffix ?? (mode === 'room' ? ' in this room' : '');
+
   return async (roomId, event) => {
     try {
       if (!event.content || event.content.msgtype !== "m.text") return;
       if (event.sender === botUserId) return;
 
-      // Only respond to messages in the configured room
-      if (roomId !== config.bot.targetRoomId) return;
+      // Bot-specific gating
+      if (mode === 'dm') {
+        let members;
+        try {
+          members = await client.getJoinedRoomMembers(roomId);
+        } catch (memberError) {
+          console.error(`Failed to get room members: ${memberError.message}`);
+          return;
+        }
 
-      console.log(`Message in room ${roomId} from ${event.sender}: ${event.content.body}`);
+        const isDM = members.length === 2;
+        if (!isDM) return;
+
+        if (!config.bot.allowedUsers.includes(event.sender)) {
+          console.log(`Ignoring message from non-whitelisted user: ${event.sender}`);
+          return;
+        }
+
+        console.log(`DM from ${event.sender}: ${event.content.body}`);
+      } else if (mode === 'room') {
+        if (roomId !== config.bot.targetRoomId) return;
+        console.log(`Message in room ${roomId} from ${event.sender}: ${event.content.body}`);
+      }
 
       const parsed = parseCommand(event.content.body);
 
@@ -38,7 +62,7 @@ export function createMessageHandler({ client, config, axios, botUserId }) {
             originalMessage: event.content.body,
             roomId: roomId,
             timestamp: new Date().toISOString(),
-            botType: 'roombot'
+            botType: botType
           }, {
             timeout: 30000,
             headers: { 'Content-Type': 'application/json' }
@@ -74,7 +98,7 @@ export function createMessageHandler({ client, config, axios, botUserId }) {
       try {
         await client.sendMessage(roomId, {
           msgtype: "m.text",
-          body: `Hello ${event.sender}, I received your message in this room: "${event.content.body}"`
+          body: `Hello ${event.sender}, I received your message${fallbackSuffix}: "${event.content.body}"`
         });
       } catch (sendError) {
         console.error(`Failed to send message to ${roomId}: ${sendError.message}`);
