+++ shared/tests/commands.test.js
